+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Unreleased
+- Fixed a bug involving the load promise of embedded collections
 
 ### 1.1.0
 - Add experimental support for using the plugin in Nuxt.js applications